+++ scout/commands/download/ensembl.py
@@ -1,29 +1,18 @@
 """Code for handling downloading of ensembl files used by scout from CLI"""
 import logging
 import pathlib
+from typing import List, Optional
 
 import click
-
-from scout.utils.scout_requests import (
-    fetch_ensembl_exons,
-    fetch_ensembl_genes,
-    fetch_ensembl_transcripts,
-)
 
 LOG = logging.getLogger(__name__)
 
 
-def print_ensembl(out_dir, resource_type, genome_build=None):
-    """Fetch and print ensembl info to file
+def print_ensembl(
+    out_dir: click.Path, resource_type: List[str], genome_build: Optional[str] = None
+):
+    """Fetch and print Ensembl info to file. If no genome build is used both builds will be fetched."""
 
-    If no genome build is used both builds will be fetched
-
-    Args:
-        out_dir(Path): Path to existing directory
-        resource_type(str): in ['genes', 'transcripts', 'exons']
-        genome_build
-
-    """
     if resource_type not in ["genes", "transcripts", "exons"]:
         LOG.error("Invalid resource type")
         raise SyntaxError()